Minimal TS test compiles
