+++ package.json
@@ -6,7 +6,7 @@
     "pnpm": ">=6"
   },
   "packageManager": "pnpm@7.1.1",
-  "workspaces": [],
+  "type": "module",
   "scripts": {
     "start": "pnpm network 2",
     "network": "hc s clean && pnpm build:happ && concurrently-repeat \"pnpm start:agent\"",
@@ -25,6 +25,8 @@
     "@holochain/tryorama": "^0.9.1",
     "@msgpack/msgpack": "^2.8.0",
     "@types/node": "^18.11.9",
+    "@types/tape": "^4.13.2",
+    "@types/tape-promise": "^4.0.1",
     "concurrently": "^6.2.1",
     "concurrently-repeat": "^0.0.1",
     "cross-env": "^7.0.3",
@@ -36,6 +38,5 @@
     "ts-node": "^10.9.1",
     "typescript": "^4.8.4",
     "ws": "^8.11.0"
-  },
-  "type": "module"
+  }
 }
